Moved internal classes into internal package
