+++ engine/src/main/java/de/devathlon/hnl/engine/GameEngine.java
@@ -5,10 +5,9 @@ import de.devathlon.hnl.core.pause.PauseItem;
 import de.devathlon.hnl.core.update.EngineUpdate;
 import de.devathlon.hnl.engine.configuration.EngineConfiguration;
 import de.devathlon.hnl.engine.listener.InputListener;
-import de.devathlon.hnl.engine.update.EffectInformation;
-import de.devathlon.hnl.engine.update.Score;
+import de.devathlon.hnl.engine.internal.update.EffectInformation;
+import de.devathlon.hnl.engine.internal.update.Score;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.function.Consumer;
