+++ engine/src/main/java/de/devathlon/hnl/engine/GameEngineImpl.java
@@ -5,11 +5,11 @@ import de.devathlon.hnl.core.pause.PauseItem;
 import de.devathlon.hnl.core.update.EngineUpdate;
 import de.devathlon.hnl.engine.configuration.EngineConfiguration;
 import de.devathlon.hnl.engine.listener.InputListener;
-import de.devathlon.hnl.engine.loop.GameLoop;
-import de.devathlon.hnl.engine.update.EffectInformation;
-import de.devathlon.hnl.engine.window.GameCanvas;
-import de.devathlon.hnl.engine.window.GameWindow;
-import de.devathlon.hnl.engine.update.Score;
+import de.devathlon.hnl.engine.internal.loop.GameLoop;
+import de.devathlon.hnl.engine.internal.update.EffectInformation;
+import de.devathlon.hnl.engine.internal.window.GameCanvas;
+import de.devathlon.hnl.engine.internal.window.GameWindow;
+import de.devathlon.hnl.engine.internal.update.Score;
 
 import java.util.Arrays;
 import java.util.LinkedList;
